Restructuring to module treelstm
