+++ treelstm/tree.py
