+++ treelstm/vocab.py
